Fix bugs for large cases (#34)
* Fix square_intersection function

* Use Float32 instead of Number/Real/Float64 for coordinate related computations in topo (also Point2f instead of Tuple and Vector). Remove redundant notify_component and observables (use the @lift macro instead).

* Add missing tests for show-function on the types AbstractSystem and ProcInvData

* Improve performance of updates to ax_info and other misc improvements

* Fix bug that made nodes/areas disappear when plotting too many objects (max z_level is 10000)

* Run tests on Julia version 1.11 due to incompatibility with version 1.12

+++ src/utils_gen/utils.jl
@@ -182,16 +182,16 @@ function get_max_installed(::Any, ::Vector{<:TS.TimeStructure})
 end
 
 """
-    mouse_within_axis(ax::Makie.AbstractAxis, mouse_pos::Tuple{Float64,Float64})
+    mouse_within_axis(ax::Makie.AbstractAxis, mouse_pos::Tuple{Float32,Float32})
 
 Check if mouse position is within the pixel area of `ax`.
 """
-function mouse_within_axis(ax::Makie.AbstractAxis, mouse_pos::Tuple{Float64,Float64})
+function mouse_within_axis(ax::Makie.AbstractAxis, mouse_pos::Tuple{Float32,Float32})
     origin::Vec2{Int64} = pixelarea(ax.scene)[].origin
     widths::Vec2{Int64} = pixelarea(ax.scene)[].widths
-    mouse_pos_loc::Vec2{Float64} = mouse_pos .- origin
+    mouse_pos_loc::Vec2{Float32} = mouse_pos .- origin
 
-    return all(mouse_pos_loc .> 0.0) && all(mouse_pos_loc .- widths .< 0.0)
+    return all(mouse_pos_loc .> 0.0f0) && all(mouse_pos_loc .- widths .< 0.0f0)
 end
 
 """
